requirements.txt
streamlit
openai>=1.3.0
pillow
pymupdf
pandas
python-dotenv